v2.7.0
+ refactored kafka stream
  - added a record iterator in addition to callback function
  - added get method templates
+ minor edits
  - changed a bug where axon::timer destructor of anonymous variable was behaving
  - fixed few issues with scylladb driver
  - fixed a memory leak issue
