+++ src/kafka.cpp
@@ -21,6 +21,255 @@ namespace axon
 {
 	namespace stream
 	{
+		subscription::subscription(int count):
+			_subscription(NULL)
+		{
+			_subscription = rd_kafka_topic_partition_list_new(count);
+		}
+
+		subscription::~subscription()
+		{
+			if (_subscription)
+				rd_kafka_topic_partition_list_destroy(_subscription);
+		}
+
+		void subscription::add(std::string topic)
+		{
+			if (rd_kafka_topic_partition_list_add(_subscription, topic.c_str(), RD_KAFKA_PARTITION_UA) == NULL)
+				std::cerr<<"null returned"<<std::endl;
+		}
+
+		int subscription::count()
+		{
+			return _subscription->cnt;
+		}
+
+		rd_kafka_topic_partition_list_t *subscription::get()
+		{
+			if (_subscription == NULL)
+				throw axon::exception(__FILENAME__, __LINE__, __PRETTY_FUNCTION__, "subscription not ready");
+			return _subscription;
+		}
+
+		void subscription::info()
+		{
+			for (int i = 0; i < _subscription->cnt; i++) {
+				rd_kafka_topic_partition_t *p = &_subscription->elems[i];
+				printf("Topic \"%s\" partition %" PRId32, p->topic,
+						p->partition);
+				if (p->err)
+						printf(" error %s", rd_kafka_err2str(p->err));
+				else {
+						printf(" offset %" PRId64 "", p->offset);
+
+						if (p->metadata_size)
+								printf(" (%d bytes of metadata)",
+										(int)p->metadata_size);
+				}
+				printf("\n");
+			}
+		}
+
+		std::string recordset::_get_string(std::string name)
+		{
+			std::string value;
+			avro_value_t init_iden_value;
+
+			if (avro_value_get_by_name(&_data, name.c_str(), &init_iden_value, NULL) == 0)
+			{
+				int retcode = 0;
+				avro_value_t branch;
+
+				if ((retcode = avro_value_set_branch(&init_iden_value, 1, &branch)) != 0)
+					throw axon::exception(__FILENAME__, __LINE__, __PRETTY_FUNCTION__, "branch not found");
+
+				if (avro_value_get_type(&branch) == AVRO_STRING)
+				{
+					const char *buffer;
+					size_t size;
+
+					avro_value_get_string(&branch, &buffer, &size);
+
+					if (size <= 0)
+						return 0;
+
+					value = buffer;
+				}
+				else
+					throw axon::exception(__FILENAME__, __LINE__, __PRETTY_FUNCTION__, "type mismatch");
+			}
+			else
+				throw axon::exception(__FILENAME__, __LINE__, __PRETTY_FUNCTION__, "value not found");
+
+			return value;
+		}
+
+		int recordset::_get_int(std::string name)
+		{
+			int32_t value = 0;
+			avro_value_t init_iden_value;
+
+			if (avro_value_get_by_name(&_data, name.c_str(), &init_iden_value, NULL) == 0)
+			{
+				int retcode = 0;
+				avro_value_t branch;
+
+				if ((retcode = avro_value_set_branch(&init_iden_value, 1, &branch)) != 0)
+					throw axon::exception(__FILENAME__, __LINE__, __PRETTY_FUNCTION__, "branch not found");
+
+				if (avro_value_get_type(&branch) == AVRO_INT32)
+					avro_value_get_int(&branch, &value);
+				else
+					throw axon::exception(__FILENAME__, __LINE__, __PRETTY_FUNCTION__, "type mismatch");
+			}
+			else
+				throw axon::exception(__FILENAME__, __LINE__, __PRETTY_FUNCTION__, "value not found");
+
+			return value;
+		}
+
+		long recordset::_get_long(std::string name)
+		{
+			int64_t value = 0;
+			avro_value_t init_iden_value;
+
+			if (avro_value_get_by_name(&_data, name.c_str(), &init_iden_value, NULL) == 0)
+			{
+				int retcode = 0;
+				avro_value_t branch;
+
+				if ((retcode = avro_value_set_branch(&init_iden_value, 1, &branch)) != 0)
+					throw axon::exception(__FILENAME__, __LINE__, __PRETTY_FUNCTION__, "branch not found");
+
+				if (avro_value_get_type(&branch) == AVRO_INT64)
+					avro_value_get_long(&branch, &value);
+				else
+					throw axon::exception(__FILENAME__, __LINE__, __PRETTY_FUNCTION__, "type mismatch");
+			}
+			else
+				throw axon::exception(__FILENAME__, __LINE__, __PRETTY_FUNCTION__, "value not found");
+
+			return value;
+		}
+
+		float recordset::_get_float(std::string name)
+		{
+			float value = 0;
+			avro_value_t init_iden_value;
+
+			if (avro_value_get_by_name(&_data, name.c_str(), &init_iden_value, NULL) == 0)
+			{
+				int retcode = 0;
+				avro_value_t branch;
+
+				if ((retcode = avro_value_set_branch(&init_iden_value, 1, &branch)) != 0)
+					throw axon::exception(__FILENAME__, __LINE__, __PRETTY_FUNCTION__, "branch not found");
+
+				if (avro_value_get_type(&branch) == AVRO_FLOAT)
+					avro_value_get_float(&branch, &value);
+				else
+					throw axon::exception(__FILENAME__, __LINE__, __PRETTY_FUNCTION__, "type mismatch");
+			}
+			else
+				throw axon::exception(__FILENAME__, __LINE__, __PRETTY_FUNCTION__, "value not found");
+
+			return value;
+		}
+
+		double recordset::_get_double(std::string name)
+		{
+			double value = 0;
+			avro_value_t init_iden_value;
+
+			if (avro_value_get_by_name(&_data, name.c_str(), &init_iden_value, NULL) == 0)
+			{
+				int retcode = 0;
+				avro_value_t branch;
+
+				if ((retcode = avro_value_set_branch(&init_iden_value, 1, &branch)) != 0)
+					throw axon::exception(__FILENAME__, __LINE__, __PRETTY_FUNCTION__, "branch not found");
+
+				if (avro_value_get_type(&branch) == AVRO_DOUBLE)
+					avro_value_get_double(&branch, &value);
+				else
+					throw axon::exception(__FILENAME__, __LINE__, __PRETTY_FUNCTION__, "type mismatch");
+			}
+			else
+				throw axon::exception(__FILENAME__, __LINE__, __PRETTY_FUNCTION__, "value not found");
+
+			return value;
+		}
+
+		size_t recordset::get(std::string name, void *value, size_t size)
+		{
+			avro_value_t init_iden_value;
+
+			if (avro_value_get_by_name(&_data, name.c_str(), &init_iden_value, NULL) == 0)
+			{
+				int retcode = 0;
+				avro_value_t branch;
+
+				if ((retcode = avro_value_set_branch(&init_iden_value, 1, &branch)) != 0)
+					throw axon::exception(__FILENAME__, __LINE__, __PRETTY_FUNCTION__, "branch not found");
+
+				if (avro_value_get_type(&branch) == AVRO_BYTES)
+				{
+					const void *buffer;
+					size_t bsz = 0;
+
+					avro_value_get_bytes(&branch, &buffer, &bsz);
+
+					if (bsz <= 0)
+						return 0;
+
+					memcpy(value, buffer, std::min(size, bsz));
+
+					return std::min(size, bsz);
+				}
+				else
+					throw axon::exception(__FILENAME__, __LINE__, __PRETTY_FUNCTION__, "type mismatch");
+			}
+			else
+				throw axon::exception(__FILENAME__, __LINE__, __PRETTY_FUNCTION__, "value not found");
+
+			return 0;
+		}
+
+		recordset::recordset(axon::stream::serdes &src, rd_kafka_message_t *rkm)
+		{
+			char error[1024];
+
+			if (rkm)
+			{
+				_empty = false;
+
+				if (serdes_deserialize_avro(src.get(), &_data, NULL, rkm->payload, rkm->len, error, 1024))
+					throw axon::exception(__FILENAME__, __LINE__, __PRETTY_FUNCTION__, error);
+
+				rd_kafka_message_destroy(rkm);
+			}
+			else
+				_empty = true;
+		}
+
+		recordset::~recordset()
+		{
+			if (!_empty)
+				avro_value_decref(&_data);
+		}
+
+		void recordset::print()
+		{
+			char *as_json;
+
+			if (avro_value_to_json(&_data, 1, &as_json))
+				throw axon::exception(__FILENAME__, __LINE__, __PRETTY_FUNCTION__, std::string("avro_to_json failed: ") + avro_strerror());
+			else {
+				printf("%15s\n", as_json);
+				free(as_json);
+			}
+		}
+
 		kafka::kafka(std::string bootstrap, std::string schema, std::string consumer): 
 			_bootstrap_hosts(bootstrap),
 			_schema_hosts(schema),
@@ -40,8 +289,12 @@ namespace axon
 			config.set("enable.auto.commit", "true");
 			config.set("auto.offset.reset", "earliest");
 			config.set("bootstrap.servers", _bootstrap_hosts);
+#if DEBUG == 1
+			config.set("debug", "topic,metadata,feature");
+			// rd_kafka_conf_set_log_cb(config, logger);
+#endif
 
-			if (!(_rk = rd_kafka_new(RD_KAFKA_CONSUMER, config.get(), error, sizeof(error))))
+			if (!(_rk = rd_kafka_new(RD_KAFKA_CONSUMER, config, error, sizeof(error))))
 				throw axon::exception(__FILENAME__, __LINE__, __PRETTY_FUNCTION__, error);
 
 			rd_kafka_poll_set_consumer(_rk);
@@ -49,7 +302,7 @@ namespace axon
 			config.reset();
 
 			_runnable = false;
-			_connected = false;
+			_subscribed = false;
 
 			_counter = 0;
 		}
@@ -57,6 +310,7 @@ namespace axon
 		kafka::~kafka()
 		{
 			stop();
+			rd_kafka_destroy(_rk);
 		}
 
 		bool kafka::add(std::string topic)
@@ -66,7 +320,7 @@ namespace axon
 			return true;
 		}
 
-		void kafka::connect()
+		void kafka::subscribe()
 		{
 			if (!_rk)
 				throw axon::exception(__FILENAME__, __LINE__, __PRETTY_FUNCTION__, "kafka connection not initialized");
@@ -74,30 +328,22 @@ namespace axon
 			if (_topic.size() < 1)
 				throw axon::exception(__FILENAME__, __LINE__, __PRETTY_FUNCTION__, "topic list is empty");
 
-			_subscription = rd_kafka_topic_partition_list_new(_topic.size());
+			subscription sb(_topic.size());
 			
 			for (std::string &topic : _topic)
-				rd_kafka_topic_partition_list_add(_subscription, topic.c_str(), RD_KAFKA_PARTITION_UA);
+				sb.add(topic.c_str());
 			
-			if ((err = rd_kafka_subscribe(_rk, _subscription)))
-			{
-				DBGPRN("=> Failed to subscribe to %d topics: %s\n", _subscription->cnt, rd_kafka_err2str(err));
+			if ((err = rd_kafka_subscribe(_rk, sb)))
+				throw axon::exception(__FILENAME__, __LINE__, __PRETTY_FUNCTION__, std::string(rd_kafka_err2str(err)));
 
-				rd_kafka_topic_partition_list_destroy(_subscription);
-				rd_kafka_destroy(_rk); // <-- this needs to go in destructor
+			DBGPRN("=> Subscribed to %d topic(s), waiting for rebalance and messages...", sb.count());
 
-				throw axon::exception(__FILENAME__, __LINE__, __PRETTY_FUNCTION__, "failed to subscribe to topic");
-			}
-
-			DBGPRN("=> Subscribed to %d topic(s), waiting for rebalance and messages...\n", _subscription->cnt);
-			rd_kafka_topic_partition_list_destroy(_subscription);
-
-			_connected = true;
+			_subscribed = true;
 		}
 
-		bool kafka::start(std::function<void(avro_value_t *)> fn)
+		bool kafka::start(std::function<void(axon::stream::recordset*)> fn)
 		{
-			if (!_connected)
+			if (!_subscribed)
 				throw axon::exception(__FILENAME__, __LINE__, __PRETTY_FUNCTION__, "not connected yet!");
 
 			f = fn;
@@ -105,45 +351,31 @@ namespace axon
 
 			_runner = std::thread([this] () {
 
-				unsigned long long counter = 0;
-
 				while (_runnable)
 				{
-					axon::timer(__PRETTY_FUNCTION__);
 					rd_kafka_message_t *rkm;
 
 					if (!(rkm = rd_kafka_consumer_poll(_rk, 300)))
-					{
-						counter = 0;
 						continue;
-					}
 
 					if (rkm->err)
 					{
-						DBGPRN("=> Consumer error: %d->%s\n", rkm->err, rd_kafka_message_errstr(rkm));
+						DBGPRN("=> Consumer error: %d->%s", rkm->err, rd_kafka_message_errstr(rkm));
 						rd_kafka_message_destroy(rkm);
 
 						continue;
 					}
 
 					try {
-						avro_value_t avro;
-						char error[1024];
-
-						if (serdes_deserialize_avro(_serdes.get(), &avro, NULL, rkm->payload, rkm->len, error, 1024))
-							throw axon::exception(__FILENAME__, __LINE__, __PRETTY_FUNCTION__, error);
-
-						f(&avro);
-						avro_value_decref(&avro);
-
+						recordset rc(_serdes, rkm);
+						f(&rc);
 					} catch(const std::bad_function_call& e) {
 						std::cout<<e.what()<<std::endl;
 					}
 
-					rd_kafka_message_destroy(rkm);
-
-					if ((counter % MIN_COMMIT_COUNT) == 0)
+					if ((_counter % MIN_COMMIT_COUNT) == 0)
 					{
+						INFPRN("%s - rd_kafka_commit() at %d", __PRETTY_FUNCTION__, _counter);
 						err = rd_kafka_commit(_rk, NULL, 0);
 						if (err)
 						{
@@ -151,7 +383,6 @@ namespace axon
 						}
 					}
 
-					counter++;
 					_counter++;
 				}
 			});
@@ -167,23 +398,37 @@ namespace axon
 			if (_runner.joinable())
 				_runner.join();
 
-			if (_connected)
+			if (_subscribed)
 			{
 				rd_kafka_unsubscribe(_rk);
-				_connected = false;
+				_subscribed = false;
 			}
 
 			if (_rk)
-			{
 				rd_kafka_consumer_close(_rk);
-				rd_kafka_destroy(_rk);
-				_rk = NULL;
-			}
 		}
 
 		unsigned long long kafka::counter()
 		{
 			return _counter;
+		}
+
+		std::unique_ptr<axon::stream::recordset> kafka::next()
+		{
+			if (_runnable)
+				throw axon::exception(__FILENAME__, __LINE__, __PRETTY_FUNCTION__, "invalid operation while callback active");
+
+			rd_kafka_message_t *rkm = rd_kafka_consumer_poll(_rk, 300);
+
+			if (rkm && rkm->err)
+			{
+				DBGPRN("=> Consumer error: %d->%s", rkm->err, rd_kafka_message_errstr(rkm));
+				rd_kafka_message_destroy(rkm);
+
+				return std::unique_ptr<axon::stream::recordset>(nullptr);
+			}
+
+			return std::make_unique<axon::stream::recordset>(_serdes, rkm);
 		}
 	}
 }
